refactored the codes
