+++ src/main.js
@@ -16,6 +16,8 @@ Vue.config.productionTip = false;
 
 axios.defaults.baseURL =
   "https://firestore.googleapis.com/v1/projects/typing-app-f08b8/databases/(default)/documents";
+axios.defaults.headers.post["Content-Type"] = "application/json;charset=utf-8";
+axios.defaults.headers.post["Access-Control-Allow-Origin"] = "*";
 
 store.dispatch("autoLogin").then(() => {
   new Vue({
