+++ src/views/About.vue
@@ -5,8 +5,6 @@
 </template>
 
 <script>
-// @ is an alias to /src
-
 export default {
   name: "About"
 };
